+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/_params.py
@@ -83,7 +83,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-statements
         c.argument('yes', options_list=['--yes', '-y'], help='Do not prompt for confirmation.', action='store_true')
         c.argument('image_names', arg_type=image_by_tag_type, action='append')
         c.argument('timeout', type=int, help='The timeout in seconds.')
-        c.argument('docker_file_path', options_list=['--file', '-f'], help="The relative path of the the docker file to the source code root folder.")
+        c.argument('docker_file_path', options_list=['--file', '-f'], help="The relative path of the the docker file to the source code root folder. Default to 'Dockerfile'.")
         c.argument('no_logs', help="Do not show logs after successfully queuing the build.", action='store_true')
         c.argument('no_wait', help="Do not wait for the run to complete and return immediately after queuing the run.", action='store_true')
         c.argument('no_format', help="Indicates whether the logs should be displayed in raw format", action='store_true')
@@ -207,6 +207,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-statements
         # Source Trigger parameters
         c.argument('source_trigger_name', help="The name of the source trigger.")
         c.argument('commit_trigger_enabled', help="Indicates whether the source control commit trigger is enabled.", arg_type=get_three_state_flag())
+        c.argument('pull_request_trigger_enabled', help="Indicates whether the source control pull request trigger is enabled.", arg_type=get_three_state_flag())
         c.argument('git_access_token', help="The access token used to access the source control provider.")
         c.argument('branch', help="The source control branch name.")
         c.argument('base_image_trigger_name', help="The name of the base image trigger.")