+++ src/command_modules/azure-cli-container/azure/cli/command_modules/container/custom.py
@@ -35,8 +35,7 @@ from azure.mgmt.containerinstance.models import (AzureFileVolume, Container, Con
                                                  GitRepoVolume, LogAnalytics, ContainerGroupDiagnostics, ContainerGroupNetworkProfile,
                                                  ContainerGroupIpAddressType, ResourceIdentityType, ContainerGroupIdentity)
 from azure.cli.core.util import sdk_no_wait
-from ._client_factory import (cf_container_groups, cf_container, cf_log_analytics, cf_resource,
-                              cf_network, get_auth_management_client)
+from ._client_factory import (cf_container_groups, cf_container, cf_log_analytics, cf_resource, cf_network)
 
 logger = get_logger(__name__)
 WINDOWS_NAME = 'Windows'
@@ -107,7 +106,6 @@ def create_container(cmd,
                      assign_identity=None,
                      identity_scope=None,
                      identity_role='Contributor',
-                     identity_role_id=None,
                      no_wait=False):
     """Create a container group. """
     if file:
@@ -223,9 +221,10 @@ def create_container(cmd,
                       name, cgroup)
 
     if assign_identity is not None and identity_scope:
+        from azure.cli.core.commands.arm import assign_identity
         cg = container_group_client.get(resource_group_name, name)
-        _create_update_msi_role_assignment(cmd, resource_group_name, name, cg.identity.principal_id,
-                                           identity_role_id, identity_scope)
+        assign_identity(cmd.cli_ctx, lambda: cg, lambda cg: cg, identity_role, identity_scope)
+
     return lro
 
 
@@ -243,43 +242,6 @@ def _build_identities_info(identities):
     if external_identities:
         identity.user_assigned_identities = {e: {} for e in external_identities}
     return identity
-
-
-def _create_update_msi_role_assignment(cmd, resource_group_name, cg_name, identity_principle_id,
-                                       role_definition_id, identity_scope):
-    from azure.cli.core.profiles import ResourceType, get_sdk
-    from msrestazure.azure_exceptions import CloudError
-
-    RoleAssignmentCreateParameters = get_sdk(cmd.cli_ctx, ResourceType.MGMT_AUTHORIZATION,
-                                             'RoleAssignmentCreateParameters',
-                                             mod='models', operation_group='role_assignments')
-
-    assignments_client = get_auth_management_client(cmd.cli_ctx, identity_scope).role_assignments
-
-    role_assignment_guid = str(_gen_guid())
-
-    parameters = RoleAssignmentCreateParameters(
-        role_definition_id=role_definition_id,
-        principal_id=identity_principle_id
-    )
-
-    logger.info("Creating an assignment with a role '%s' on the scope of '%s'", role_definition_id, identity_scope)
-    retry_times = 30
-    for l in range(0, retry_times):
-        try:
-            assignments_client.create(scope=identity_scope, role_assignment_name=role_assignment_guid,
-                                      parameters=parameters)
-            break
-        except CloudError as ex:
-            if 'role assignment already exists' in ex.message:
-                logger.info('Role assignment already exists')
-                break
-            elif l < retry_times and ' does not exist in the directory ' in ex.message:
-                time.sleep(5)
-                logger.warning('Retrying role assignment creation: %s/%s', l + 1, retry_times)
-                continue
-            else:
-                raise
 
 
 def _get_resource(client, resource_group_name, *subresources):