+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/_validators.py
@@ -244,6 +244,8 @@ def validate_er_peer_circuit(cmd, namespace):
             name=namespace.peer_circuit,
             child_type_1='peerings',
             child_name_1=namespace.peering_name)
+    else:
+        peer_id = namespace.peer_circuit
 
     # if the circuit ID is provided, we need to append /peerings/{peering_name}
     if namespace.peering_name not in peer_id:
@@ -711,6 +713,7 @@ def process_nic_create_namespace(cmd, namespace):
 def process_public_ip_create_namespace(cmd, namespace):
     get_default_location_from_resource_group(cmd, namespace)
     validate_public_ip_prefix(cmd, namespace)
+    validate_ip_tags(cmd, namespace)
     validate_tags(namespace)