+++ src/command_modules/azure-cli-profile/azure/cli/command_modules/profile/__init__.py
@@ -24,6 +24,7 @@ class ProfileCommandsLoader(AzCommandsLoader):
         with self.command_group('', profile_custom) as g:
             g.command('login', 'login')
             g.command('logout', 'logout')
+            g.command('self-test', 'check_cli', deprecate_info=g.deprecate(hide=True))
 
         with self.command_group('account', profile_custom) as g:
             g.command('list', 'list_subscriptions', table_transformer=transform_account_list)
@@ -50,6 +51,7 @@ class ProfileCommandsLoader(AzCommandsLoader):
             c.argument('identity_port', type=int, help="the port to retrieve tokens for login. Default: 50342", arg_group='Managed Service Identity')
             c.argument('use_device_code', action='store_true',
                        help="Use CLI's old authentication flow based on device code. CLI will also use this if it can't launch a browser in your behalf, e.g. in remote SSH or Cloud Shell")
+            c.argument('use_cert_sn_issuer', action='store_true', help='used with a service principal configured with Subject Name and Issuer Authentication in order to support automatic certificate rolls')
 
         with self.argument_context('logout') as c:
             c.argument('username', help='account user, if missing, logout the current active account')