+++ src/azure-cli-core/HISTORY.rst
@@ -2,10 +2,9 @@
 
 Release History
 ===============
-
 2.0.50
 ++++++
-* Minor fixes.
+* auth: support service principal sn+issuer auth
 
 2.0.49
 ++++++